src/requirements.txt
# Scraping
requests

# API clients
openai
aiohttp
apify-client

# Environment and configuration
python-dotenv
backports.zoneinfo;python_version<"3.9"

# System monitoring
psutil

# Task scheduling
APScheduler

# Messaging platforms
python-telegram-bot[job-queue]
slack-bolt
